feat(dashboard): complete web dashboard with Discord OAuth2, SaaS pricing tiers, admin portal, and full PostgreSQL migration

+++ internal/storage/migrations/003_add_whitelist.sql
@@ -1,13 +1,13 @@
 CREATE TABLE IF NOT EXISTS whitelist_users (
-  guild_id TEXT NOT NULL,
-  user_id TEXT NOT NULL,
-  created_at INTEGER NOT NULL,
-  PRIMARY KEY (guild_id, user_id)
+    guild_id TEXT NOT NULL,
+    user_id TEXT NOT NULL,
+    created_at BIGINT NOT NULL,
+    PRIMARY KEY (guild_id, user_id)
 );
 
 CREATE TABLE IF NOT EXISTS whitelist_roles (
-  guild_id TEXT NOT NULL,
-  role_id TEXT NOT NULL,
-  created_at INTEGER NOT NULL,
-  PRIMARY KEY (guild_id, role_id)
+    guild_id TEXT NOT NULL,
+    role_id TEXT NOT NULL,
+    created_at BIGINT NOT NULL,
+    PRIMARY KEY (guild_id, role_id)
 );
